requirements.txt
# GRC Zero Trust Platform Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Redis for caching and sessions
redis>=5.0.0
aioredis>=2.0.1

# SAP Connectivity
pyrfc>=3.0.0  # Requires SAP NW RFC SDK installed

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0

# Data Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async Support
httpx>=0.25.0
aiohttp>=3.9.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Machine Learning (for risk scoring)
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0

# Graph Analysis (for SoD rule relationships)
networkx>=3.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
